requirements.txt
python-telegram-bot[job-queue]>=21.0
google-api-python-client>=2.100.0
youtube-transcript-api>=1.0.0
python-dotenv>=1.0.0
